retrieval-agent/requirements.txt
langchain
langgraph
openai
faiss-cpu
fastapi
uvicorn
python-dotenv
tqdm